refactor: member 도메인 클래스 model 패키지에서 보관
