+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/member/infrastructure/GatheringMemberRepository.java
@@ -1,7 +1,7 @@
 package com.triplem.momoim.core.domain.member.infrastructure;
 
-import com.triplem.momoim.core.domain.member.GatheringMember;
 import com.triplem.momoim.core.domain.member.dto.GatheringMemberDetail;
+import com.triplem.momoim.core.domain.member.model.GatheringMember;
 import java.util.List;
 
 public interface GatheringMemberRepository {
